move chunking to modeling_utils
